Major task2.c changes
Stuff in comments is to show work done/stuff tested. Everything un-commented is the final code so far. Code size only increased by a bit under 200 bytes but time saved is almost 0.4 seconds. What helped the time go faster is unrolling the loop one more time further, checking if we multiply by 0 and to ignore the calculation if we do, and checking if the halfway mark has to 0's in a row which means that anything greater than sample[limit] is also zero, so only compute from 0 to limit.