Updated Task3
Updated the main task3.c to include the new unrolled version (I like
btw, nice job) and the updated task2.c main program. I ran into a slight
problem that I'm not sure how to solve (see comments). That's a lie, I
know how to solve it; but, it might not be a valid solution. Might have
to ask Fabio on monday, unless I can think of a work around.