Final Changes? (Check plz)
So I fixed up some stuff but looking at the assembly there is still some stuff I don't like but having a hard time fixing. Anyways its still really good. Program Size: Code=2516 RO-data=32 RW-data=14612 ZI-data=2284 and t = 2.83829367 (including startup and everything until the return statement on line 57).
Doing the int version of this is just a port, of which I saw you did it already, but i'll change some of the small tweaks I made.
Fun fact: part 1 is pretty slow, but thats because this is optimized for part 2. I can explain better in person on monday. But try it out if you want.