Final version (hopefully)
So just copy and pasted the function from part 2. The reason this is not as fast is that this part is only to test R[0]-R[255] vs Coeff[0]-Coeff[256], but our code optimizes alot for zeros, which in this test case there is none (or very little). t = 0.02314942 for the whole code up to the return 0 (line 44) and Program Size: Code=1932 RO-data=32 RW-data=14600 ZI-data=1256