Chyziak Update
So I guess u just missed that for this version the threshold changes to 0x69000 haha. So it works. I also changed it to the new version of the function like in part 2 and 1, but I copied the rest from yours and it works (you missed some numbers that were suppose to be ints and kept as floats but thats just silly rushing mistake probably). 
Not sure if there is some way we can optimize the int version that the float couldnt be optimized. So have a look and try and pry it apart, thats what I'm gonna do and if I get no idea tonight i'll just give up and say its good enough, cause damn its pretty good. (I want sub 1 sec tho :D).
t = 1.15583483 and Program Size: Code=1412 RO-data=32 RW-data=14612 ZI-data=2284