Updated task4.c
Coded the loop, and cleaned up the code. I can't get it to work out to a
peak at 715. Without unrolling, the peak is at 812, unrolling it goes to
255. I'm certain it's a mistake in my logic, but I think I've been
staring at it too long to see it.

I also backed up the messy code I had to task4_backup1.c. This file
contains all of the task 3 code, commented out.